move request logger to /middleware
